URI segments can now be passed into route callbacks as parameters.
Examples:
```
Router::get('/article/{author}/{article_id}', function ($author, $articleID) {
	return "Welcome to article $articleID written by $author";
});
```

When the above route is visited as `/article/tim/256/` you will see 'Welcome to article 256 written by tim.' displayed
in the browser.
Note the order in which the parameters appear in the URI affects the order in which they are passed into the callback.
Non-positional arguments are not supported yet.